Added some error handling

+++ ankiAddon/__init__.py
@@ -3,7 +3,6 @@ from aqt import mw
 from anki.hooks import addHook
 from anki.utils import checksum
 from subprocess import call
-import time
 import re, os, shutil, time
 
 mypath = os.path.dirname(os.path.abspath(__file__))
@@ -13,10 +12,13 @@ def drawingPad(editor):
     drawing = os.path.join(mypath, fname)
     call(["java", "-jar", os.path.join(mypath, "drawingPad.jar"), drawing])
     
-    mdir = mw.col.media.dir()
-    shutil.copyfile(drawing, os.path.join(mdir, fname))
-    editor.web.eval("wrap(\"<img src='"+fname+"' />\", \"\");")
-    os.remove(drawing)
+    if os.path.exists(drawing):
+        mdir = mw.col.media.dir()
+        shutil.copyfile(drawing, os.path.join(mdir, fname))
+        editor.web.eval("wrap(\"<img src='"+fname+"' />\", \"\");")
+        os.remove(drawing)
+    else:
+        editor.web.eval("wrap(\"Unable to start DrawingPad. Do you have JavaFx installed?\", \"\");")
 
 def addMyButton(buttons, editor):
     editor._links['drawingPad'] = drawingPad
